[WFLY-11440] Refactoring - removing deprecated EJBUtilities class.
