+++ ejb3/src/main/java/org/jboss/as/ejb3/component/EJBComponent.java
@@ -63,6 +63,7 @@ import org.jboss.as.ejb3.context.CurrentInvocationContext;
 import org.jboss.as.ejb3.logging.EjbLogger;
 import org.jboss.as.ejb3.security.EJBSecurityMetaData;
 import org.jboss.as.ejb3.security.JaccInterceptor;
+import org.jboss.as.ejb3.subsystem.EJBStatistics;
 import org.jboss.as.ejb3.suspend.EJBSuspendHandlerService;
 import org.jboss.as.ejb3.timerservice.TimerServiceImpl;
 import org.jboss.as.ejb3.tx.ApplicationExceptionDetails;
@@ -89,6 +90,7 @@ import org.wildfly.transaction.client.ContextTransactionSynchronizationRegistry;
 
 /**
  * @author <a href="mailto:cdewolf@redhat.com">Carlo de Wolf</a>
+ * @author <a href="mailto:ropalka@redhat.com">Richard Opalka</a>
  */
 public abstract class EJBComponent extends BasicComponent implements ServerActivityCallback {
 
@@ -98,7 +100,6 @@ public abstract class EJBComponent extends BasicComponent implements ServerActiv
     private final Map<MethodTransactionAttributeKey, Integer> txTimeouts;
     private final Map<MethodTransactionAttributeKey, Boolean> txExplicitAttrs;
 
-    private final EJBUtilities utilities;
     private final boolean isBeanManagedTransaction;
     private final Map<Class<?>, ApplicationExceptionDetails> applicationExceptions;
     private final EJBSecurityMetaData securityMetaData;
@@ -145,11 +146,7 @@ public abstract class EJBComponent extends BasicComponent implements ServerActiv
      */
     protected EJBComponent(final EJBComponentCreateService ejbComponentCreateService) {
         super(ejbComponentCreateService);
-
-
         this.applicationExceptions = Collections.unmodifiableMap(ejbComponentCreateService.getApplicationExceptions().getApplicationExceptions());
-
-        this.utilities = ejbComponentCreateService.getEJBUtilities();
         final Map<MethodTransactionAttributeKey, TransactionAttributeType> txAttrs = ejbComponentCreateService.getTxAttrs();
         if (txAttrs == null || txAttrs.isEmpty()) {
             this.txAttrs = Collections.emptyMap();
@@ -468,7 +465,7 @@ public abstract class EJBComponent extends BasicComponent implements ServerActiv
     }
 
     public boolean isStatisticsEnabled() {
-        return utilities.isStatisticsEnabled();
+        return EJBStatistics.getInstance().isEnabled();
     }
 
     public Object lookup(String name) throws IllegalArgumentException {
