+++ ejb3/src/main/java/org/jboss/as/ejb3/subsystem/EJB3SubsystemAdd.java
@@ -39,7 +39,6 @@ import static org.jboss.as.ejb3.subsystem.EJB3SubsystemRootResourceDefinition.EJ
 import java.net.URI;
 import java.util.function.Supplier;
 
-import org.jboss.as.connector.util.ConnectorServices;
 import org.jboss.as.controller.AbstractBoottimeAddStepHandler;
 import org.jboss.as.controller.AttributeDefinition;
 import org.jboss.as.controller.OperationContext;
@@ -49,7 +48,6 @@ import org.jboss.as.controller.ProcessType;
 import org.jboss.as.controller.RunningMode;
 import org.jboss.as.controller.registry.Resource;
 import org.jboss.as.ejb3.clustering.SingletonBarrierService;
-import org.jboss.as.ejb3.component.EJBUtilities;
 import org.jboss.as.ejb3.deployment.DeploymentRepository;
 import org.jboss.as.ejb3.deployment.DeploymentRepositoryService;
 import org.jboss.as.ejb3.deployment.processors.AnnotatedEJBComponentDescriptionDeploymentUnitProcessor;
@@ -135,7 +133,6 @@ import org.jboss.as.txn.service.UserTransactionAccessControlService;
 import org.jboss.dmr.ModelNode;
 import org.jboss.ejb.client.EJBTransportProvider;
 import org.jboss.javax.rmi.RemoteObjectSubstitutionManager;
-import org.jboss.jca.core.spi.rar.ResourceAdapterRepository;
 import org.jboss.metadata.ejb.spec.EjbJarMetaData;
 import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceController;
@@ -156,6 +153,7 @@ import io.undertow.server.handlers.PathHandler;
  * Add operation handler for the EJB3 subsystem.
  *
  * @author Emanuel Muckenhuber
+ * @author <a href="mailto:ropalka@redhat.com">Richard Opalka</a>
  */
 class EJB3SubsystemAdd extends AbstractBoottimeAddStepHandler {
 
@@ -414,13 +412,6 @@ class EJB3SubsystemAdd extends AbstractBoottimeAddStepHandler {
                 .install();
 
         if (!appclient) {
-            final EJBUtilities utilities = new EJBUtilities();
-            ServiceBuilder<EJBUtilities> ejbUtilsBuilder = serviceTarget.addService(EJBUtilities.SERVICE_NAME, utilities)
-                    .addDependency(ConnectorServices.RA_REPOSITORY_SERVICE, ResourceAdapterRepository.class, utilities.getResourceAdapterRepositoryInjector())
-                    .setInitialMode(ServiceController.Mode.PASSIVE);
-            ejbUtilsBuilder.install();
-
-
             // create the POA Registry use by iiop
             final POARegistry poaRegistry = new POARegistry();
             context.getServiceTarget().addService(POARegistry.SERVICE_NAME, poaRegistry)
