+++ ejb3/src/main/java/org/jboss/as/ejb3/subsystem/EJBStatistics.java
@@ -1,0 +1,47 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright (c) 2018, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+
+package org.jboss.as.ejb3.subsystem;
+
+/**
+ * @author <a href="mailto:ropalka@redhat.com">Richard Opalka</a>
+ */
+public final class EJBStatistics {
+
+    private static final EJBStatistics INSTANCE = new EJBStatistics();
+    private volatile boolean enabled;
+
+    private EJBStatistics() {}
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    void setEnabled(final boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public static EJBStatistics getInstance() {
+        return INSTANCE;
+    }
+
+}
